--- base/rico-core/src/main/java/dev/rico/core/trace/Span.java
@@ -1,13 +1,0 @@
-package dev.rico.core.trace;
-
-import dev.rico.core.context.Context;
-
-public interface Span {
-
-    void complete();
-
-    void completeExceptional(Throwable e);
-
-    void addContext(Context context);
-
-}
