--- base/rico-core/src/main/java/dev/rico/core/trace/SpanType.java
@@ -1,5 +1,0 @@
-package dev.rico.core.trace;
-
-public enum SpanType {
-    LOCAL, SERVICE, CLIENT, UNDEFINED;
-}
