--- base/rico-core/src/main/java/dev/rico/internal/core/tracing/SpanImpl.java
@@ -1,46 +1,0 @@
-package dev.rico.internal.core.tracing;
-
-import dev.rico.core.context.Context;
-import dev.rico.core.trace.Span;
-import dev.rico.internal.core.Assert;
-
-import java.util.Objects;
-
-public class SpanImpl implements Span {
-
-    private final brave.Span innerSpan;
-
-    public SpanImpl(final brave.Span innerSpan) {
-        this.innerSpan = Assert.requireNonNull(innerSpan, "innerSpan");
-    }
-
-    @Override
-    public void complete() {
-        innerSpan.finish();
-    }
-
-    @Override
-    public void completeExceptional(final Throwable exception) {
-        Assert.requireNonNull(exception, "exception");
-        innerSpan.error(exception);
-    }
-
-    @Override
-    public void addContext(final Context context) {
-        Assert.requireNonNull(context, "context");
-        innerSpan.tag(context.getType(), context.getValue());
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        final SpanImpl span = (SpanImpl) o;
-        return Objects.equals(innerSpan, span.innerSpan);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(innerSpan);
-    }
-}
