+++ incubation/rico-tracing-core/gradle.properties
@@ -1,0 +1,1 @@
+publishJars = true
