+++ incubation/rico-tracing-core/src/main/java/dev/rico/tracing/Span.java
@@ -1,0 +1,13 @@
+package dev.rico.tracing;
+
+import dev.rico.core.context.Context;
+
+public interface Span {
+
+    void complete();
+
+    void completeExceptional(Throwable e);
+
+    void addContext(Context context);
+
+}
