+++ incubation/rico-tracing-core/src/main/java/dev/rico/tracing/SpanType.java
@@ -1,0 +1,5 @@
+package dev.rico.tracing;
+
+public enum SpanType {
+    LOCAL, SERVICE, CLIENT, UNDEFINED;
+}
