+++ incubation/rico-tracing-server/gradle.properties
@@ -1,0 +1,1 @@
+publishJars = true
